中途半端だった電卓機能をリセットし、コードの整理をしました。
