--- lib/calculation.dart
@@ -1,15 +1,0 @@
-class Calculator{
-
-    var number = new List();
-    var op = new List();
-
-  static void GetKey(String letter){
-    //　TODO キー入力　-> 配列に格納　
-  }
-
-  static String Execute(){
-    int result;
-    // TODO 計算処理
-    return result.toString();
-  }
-}
