+++ lib/text_data.dart
@@ -1,0 +1,5 @@
+class TextData{
+
+  static const String appTitle = 'msymCalclator';
+
+}
